- approved cell tooltip - warning message when excel is opened - while in source term edit and hit tab you will be moved to the target area - source term edit by double click/F2

+++ Sdl.Community.ExcelTerminology/Sdl.Community.ExcelTerminology/Services/ExcelTermProviderService.cs
@@ -57,7 +57,7 @@ namespace Sdl.Community.ExcelTerminology.Services
 
         public async Task DeleteEntry(int entryId)
         {
-
+            if (entryId == 0) return;
             await _excelTermLoaderService.DeleteTerm(entryId);
 
         }
